style: 💄 Applay clang-format to runtime
1. applay clang
2. re-generate if for loop's test_case

+++ examples/runtime/simpleMain.cpp
@@ -1,7 +1,7 @@
+#include <glog/logging.h>
 #include <cstdio>
 #include <cstring>
 #include <fstream>
-#include <glog/logging.h>
 #include <iostream>
 #include <memory>
 #include <set>
@@ -11,7 +11,8 @@
 using namespace nnrt;
 
 
-torch::Tensor load_tensor(const std::string& bin_file, const std::vector<int64_t>& shape, DataType dtype) {
+torch::Tensor load_tensor(const std::string& bin_file, const std::vector<int64_t>& shape, DataType dtype) 
+{
     // ref: // https://www.cnblogs.com/yaos/p/12105108.html
     // Get num-elements
     int num_elements = 1;
@@ -66,34 +67,38 @@ void run_rnnt_from_file(std::string ir_file) {
     auto output_tensors = runtime.inferenceModel(input_tensors);
 }
 
-int main(int argc, const char* argv[]) {
+int main(int argc, const char* argv[])
+{
     LOG(INFO) << "start runtime! ";
 
-    auto print_error = [](){
+    auto print_error = []() {
         LOG(ERROR) << "Usage: ./simpleMian model_path! or set ENV (export GRAPH_IR_FILE=path/to/your/graph_ir)";
         return -1;
     };
 
     char* ir_file = nullptr;
-    if(argc == 1) {
+    if (argc == 1) {
         ir_file = getenv("GRAPH_IR_FILE");
-        if(ir_file == nullptr) {
+        if (ir_file == nullptr) {
             return print_error();
         }
     } else if (argc == 2) {
         ir_file = const_cast<char*>(argv[1]);
     } else {
-         return print_error();
+        return print_error();
     }
 
     NNRuntime runtime(ir_file);
-    runtime.test(); 
-    
+    runtime.test();
+
     std::vector<torch::Tensor> input_tensors;
 
     // Set test inputs
+    // input_tensors.push_back(torch::tensor({{1,2,3},{4,5,6}}));
+    // input_tensors.push_back(torch::tensor({{20,2,30},{-4,50,70}}));
+
     // For test RNNT with one sample
-    input_tensors.push_back(torch::ones({341, 1, 240}, torch::kHalf).cuda());
+    input_tensors.push_back(torch::ones({341, 1, 240}, torch::kFloat).cuda());
     input_tensors.push_back(torch::tensor({341}, torch::kInt64));
 
     // Inference
